ci: use self-hosted runner (#4378)
* ci: use self-hosted runners in `ci` workflow


* fix(problem-273/c): fix function declaration without prototype


* test(problem-811): adjust output in `example_2`


* test(problem-1701/c): fix return type in interface


* ci: checkout without cleaning workspace


---------

+++ old-problems/1701/c/interface.cpp
@@ -5,7 +5,7 @@ double averageWaitingTime(
     int** customers, int customersSize, int* customersColSize);
 }
 
-int solution_c(std::vector<std::vector<int>> customers) {
+double solution_c(std::vector<std::vector<int>> customers) {
   std::vector<int*> customersDatas(customers.size());
   std::vector<int> customersSizes(customers.size());
   for (int i = customers.size() - 1; i >= 0; --i) {
